+++ app/jobs/expire_further_information_request_job.rb
@@ -2,6 +2,6 @@
 
 class ExpireFurtherInformationRequestJob < ApplicationJob
   def perform(further_information_request:)
-    FurtherInformationRequestExpirer.call(further_information_request:)
+    ExpireRequestable.call(requestable: further_information_request)
   end
 end
