+++ app/models/further_information_request.rb
@@ -42,4 +42,13 @@ class FurtherInformationRequest < ApplicationRecord
       6.weeks
     end
   end
+
+  def after_expired(user:)
+    update!(
+      failure_assessor_note: "Further information not supplied by deadline",
+      passed: false,
+    )
+
+    DeclineQTS.call(application_form:, user:)
+  end
 end
